+++ CHANGELOG.md
@@ -100,6 +100,10 @@
       - **Feature:** Add fields `Labels` and `Maintenance` to model struct `Cluster` and `CreateOrUpdateClusterPayload`
       - **Feature:** Add field `ApplicationLoadBalancer` to model struct `Extension`
       - **Feature:** Add field `Cni` to model struct `Network`
+- `sqlserverflex`:
+  - [v1.16.2](services/sqlserverflex/CHANGELOG.md#v1162)
+    - `v3api`:
+      - **Bugfix:** add missing `DeleteUserWaitHandler` and `CreateUserWaitHandler`
 - `telemetrylink`:
   - [v0.5.0](services/telemetrylink/CHANGELOG.md#v050)
     - **Improvement:** Add validation for `Description` field